refactor(types): tooltip
